+++ CHANGELOG.md
@@ -39,6 +39,7 @@ While we want to fix this by offering, in the future, an explicit API that allow
 - console: data sidebar bug fixes and improvements (#921)
 - cli: fix seeds incorrectly being applied to databases in config v3 (#6683)
 - cli: add `--all-databases` flag for `migrate apply`, this allows applying migrations on all connected databases in one go
+- cli-migrations: add config v3 image
 - docs: add Hasura v2 upgrade guide (#1030)
 
 ## v2.0.0-alpha.6